+++ config/aws_account/main.tf
@@ -11,10 +11,11 @@ resource "aws_config_retention_configuration" "retention" {
 }
 
 module "config_recorder" {
-  source                      = "../../config/recorder"
-  record_iam_role_arn_default = module.service_role.data["config"].role_arn
+  source                          = "../../config/recorder"
+  name_include_app_fields_default = false
+  record_iam_role_arn_default     = module.service_role.data["config"].role_arn
   record_map = {
-    account_record = {}
+    default = {}
   }
   record_s3_bucket_key_default = var.s3_bucket_key
   s3_data_map                  = var.s3_data_map