+++ config/recorder/main.tf
@@ -46,16 +46,15 @@ resource "aws_config_configuration_recorder_status" "this_status" {
 }
 
 module "athena_db" {
-  source               = "../../athena/database"
-  db_map               = local.create_query_table_map
-  name_prepend_default = "config"
-  std_map              = var.std_map
+  source  = "../../athena/database"
+  db_map  = local.create_db_x_map
+  std_map = var.std_map
 }
 
 module "athena_query_table" {
   source               = "../../athena/named_query"
   athena_database_map  = module.athena_db.data
-  name_prepend_default = "config"
+  name_prepend_default = "config_table"
   query_map            = local.create_query_table_map
   std_map              = var.std_map
 }
@@ -63,7 +62,7 @@ module "athena_query_table" {
 module "athena_query_resource_type" {
   source               = "../../athena/named_query"
   athena_database_map  = module.athena_db.data
-  name_prepend_default = "config"
+  name_prepend_default = "config_resource"
   query_map            = local.create_query_type_map
   std_map              = var.std_map
 }
@@ -71,7 +70,7 @@ module "athena_query_resource_type" {
 module "athena_query_resource_update" {
   source               = "../../athena/named_query"
   athena_database_map  = module.athena_db.data
-  name_prepend_default = "config"
+  name_prepend_default = "config_update"
   query_map            = local.create_query_update_map
   std_map              = var.std_map
 }